perf: fuse geom interp, fused E-field FFT, float32 scatter, block_until_ready

+++ gyrojax/interpolation/scatter_gather_fa.py
@@ -85,12 +85,11 @@ def scatter_to_grid_fa(
     i0, i1, j0, j1, k0, k1, wp, wt, wa = _trilinear_weights_fa(
         state.r, state.theta, state.zeta, geom, grid_shape
     )
-    # Use float64 accumulator to avoid precision loss with large N_particles.
-    # float32 has ~7 significant digits; 1M particles × weight~0.1 per cell
-    # accumulates ~1e4 adds per cell, losing ~4 digits of precision.
-    val  = state.weight.astype(jnp.float64)
+    # Use float32 accumulator — sufficient precision for typical N_particles (50k–400k)
+    # with weights ~O(1). float32 gives ~7 significant digits.
+    val  = state.weight.astype(jnp.float32)
     size = Npsi * Ntheta * Nalpha
-    grid = jnp.zeros(size, dtype=jnp.float64)
+    grid = jnp.zeros(size, dtype=jnp.float32)
 
     def add_corner(g, ii, jj, kk, w):
         flat = ii * (Ntheta * Nalpha) + jj * Nalpha + kk
@@ -107,7 +106,7 @@ def scatter_to_grid_fa(
 
     n_particles = state.weight.shape[0]
     cell_vol = (geom.psi_grid[-1] - geom.psi_grid[0]) * 2*jnp.pi * 2*jnp.pi / size
-    delta_n = (grid.reshape(grid_shape) / (n_particles * cell_vol + 1e-30)).astype(jnp.float32)
+    delta_n = grid.reshape(grid_shape) / (n_particles * cell_vol + 1e-30)
     return delta_n
